705py-main/pyproject.toml
[project]
name = "705py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "getstream-plugins-deepgram>=0.1.1",
    "getstream[webrtc]==2.3.0a5",
    "python-dotenv>=1.1.1",
    "certifi>=2023.11.17",
    "pip-system-certs>=4.0",
    "vadersentiment>=3.3.2",
    "snownlp>=0.12.3",
    "langdetect>=1.0.9",
    "pandas>=2.3.3",
    "sentence-transformers>=5.1.1",
    "numpy>=2.2.6",
    "deepmultilingualpunctuation>=1.0.1",
    "rapidfuzz>=3.14.1",
]
